Feature/student put delete

+++ backend/src/main/java/co/edu/cue/proyectonuclear/infrastructure/dao/impl/StudentDAOImpl.java
@@ -1,18 +1,26 @@
 package co.edu.cue.proyectonuclear.infrastructure.dao.impl;
 
+import co.edu.cue.proyectonuclear.domain.entities.Professor;
 import co.edu.cue.proyectonuclear.domain.entities.Student;
+import co.edu.cue.proyectonuclear.domain.entities.Subject;
+import co.edu.cue.proyectonuclear.exceptions.StudentException;
+import co.edu.cue.proyectonuclear.exceptions.SubjectException;
 import co.edu.cue.proyectonuclear.infrastructure.dao.StudentDAO;
 import co.edu.cue.proyectonuclear.mapping.dtos.CreateStudentRequestDTO;
+import co.edu.cue.proyectonuclear.mapping.dtos.ProfessorDTO;
 import co.edu.cue.proyectonuclear.mapping.dtos.StudentDTO;
 import co.edu.cue.proyectonuclear.mapping.mappers.StudentMapper;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.NoResultException;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.Query;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 @Transactional
@@ -24,9 +32,17 @@ public class StudentDAOImpl implements StudentDAO {
     StudentMapper studentMapper;
 
     @Override
-    public StudentDTO getStudentById(Long id) {
-        Student student = entityManager.find(Student.class,id);
-        return studentMapper.mapFromEntity(student);
+    public Optional<StudentDTO> getStudentById(Long id) {
+        String query = "SELECT s.* FROM student s INNER JOIN user u ON s.id = u.id WHERE u.nid = :nidProfessor";
+        Query nativeQuery = entityManager.createNativeQuery(query);
+        nativeQuery.setParameter("nid", id);
+        try{
+            Student student = (Student) nativeQuery.getSingleResult();
+            StudentDTO studentDTO = studentMapper.mapFromEntity(student);
+            return Optional.of(studentDTO);
+        }catch (NoResultException ex){
+            return Optional.empty();
+        }
     }
 
     @Override // El DAO recibe el DTO para crear el student y lo mapea y lo guarda en la base de datos para luego hacer otro mappeo de otro DTO como respuesta.
@@ -54,7 +70,15 @@ public class StudentDAOImpl implements StudentDAO {
 
     @Override
     public StudentDTO updateStudent(StudentDTO studentDTO) {
-        Student studentEntity = studentMapper.mapFromDTO(studentDTO);
+        Student studentEntity = entityManager.find(Student.class,studentDTO.id());
+        if(studentEntity == null) throw new StudentException("Can not update, the id:" + studentDTO.id() + " does not exists", HttpStatus.BAD_REQUEST);
+        studentEntity.setEmail(studentDTO.email());
+        studentEntity.setName(studentDTO.name());
+        studentEntity.setPassword(studentDTO.password());
+        studentEntity.setLastName(studentDTO.lastName());
+        studentEntity.setCareer(studentDTO.career());
+        studentEntity.setSemester(studentDTO.semester());
+        studentEntity.setSubjects(studentDTO.subjects());
         Student studentUpdated = entityManager.merge(studentEntity);
         return studentMapper.mapFromEntity(studentUpdated);
     }
@@ -62,6 +86,7 @@ public class StudentDAOImpl implements StudentDAO {
     @Override
     public StudentDTO deleteStudent(Long id) {
         Student studentEntity = entityManager.find(Student.class,id);
+        if(studentEntity == null) throw new  SubjectException("Can not delete, the id:" + id + " does not exists", HttpStatus.BAD_REQUEST);
         entityManager.remove(studentEntity);
         return studentMapper.mapFromEntity(studentEntity);
     }
